requirements.txt
boto3
botocore
langchain-aws
ipywidgets
langgraph
chromadb
typing-extensions
langchain-community
langchain-core
langchain_huggingface
langchain
langchain_aws
langgraph
duckduckgo-search
tavily-python
crewai
grpcio==1.48 
grpcio-tools==1.48
python-dotenv
setuptools==69.5.1
crewai[tools]
unstructured
sec_api
opensearch-py
retrying
#pygraphviz
